meiduo_mall/requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.5
django-celery-results==1.2.1
django-cors-headers==3.2.1
django-crispy-forms==1.9.1
django-crontab==0.7.1
django-extensions==2.2.9
django-filter==2.2.0
django-formtools==2.2
django-haystack==2.8.1
django-import-export==2.1.0
django-redis==4.11.0
django-reversion==3.0.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
dnspython==1.16.0
elasticsearch==2.4.1
et-xmlfile==1.0.1
eventlet==0.25.2
future==0.18.2
greenlet==0.4.15
httplib2==0.9.2
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
jdcal==1.4.1
kombu==4.6.8
MarkupPy==1.14
monotonic==1.5
mutagen==1.44.0
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==2.6.4
Pillow==7.1.1
pycryptodomex==3.9.7
PyJWT==1.7.1
python-alipay-sdk==1.8.0
pytz==2019.3
PyYAML==5.3.1
QQLoginTool==0.3.0
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
uWSGI==2.0.18
vine==1.3.0
zipp==3.1.0
